# performance improvement to linearity check for rush-larsen using caching