--- a/chaste_codegen/_linearity_check.py
+++ b/chaste_codegen/_linearity_check.py
@@ -1,5 +1,5 @@
 from enum import Enum
-from functools import lru_cache
+#from functools import lru_cache
 
 from cellmlmanip.model import Variable
 from sympy import (