## Description
<!--- Describe your changes in detail -->
Slight performance update for rush-larsen models using caching in the linearity check

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here, using the 'fixes #<issue>' syntax. -->
It improves performance

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Documentation checklist
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
- [x] I have updated all documentation in the code where necessary.
- [x] I have checked spelling in all (new) comments and documentation.
- [x] I have added a note to RELEASE.md if relevant (new feature, breaking change, or notable bug fix).
- [ ] I have updated version & citation.txt & citation.cff version.

## Testing
- [X] Testing is done automatically and codecov shows test coverage
- [ ] This cannot be tested automatically <!-- describe how it has been tested)-->

I tried with several models on my laptop and it's faster:
| Model   (rush-larsen)|      Time without      |  Time With |
|----------|:----------------------:|-------------:|
| shannon_wang_puglisi_weber_bers_2004 | 25s | 25s |
| ten_tusscher_model_2006_epi| 11s | 11s |
| mahajan_shiferaw_2008| 18s | 16s |
| hund_rudy_2004| 28s | 26s |
| grandi_pasqualini_bers_2010_ss| 20s | 18s |
| ohara_rudy_2011_endo| 57s | 21s |
| paci_hyttinen_aaltosetala_severi_ventricularVersion| 13s | 12s |
| ohara_rudy_cipa_v1_2017| 40s | 19s |